pyproject.toml
[tool.poetry]
name = "y"
version = "0.1.0"
description = "A simple backend app made using Django, DjangoRestFramework and MongoDb"
authors = ["SirAbhi13 <abhinav.sny.2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
django = "3.2"
djangorestframework = "3.14"
pymongo = "^4.6.0"
pre-commit = "^3.5.0"
firebase-admin = "^6.2.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
